move snapshot code as well.
